Add testing with MySQL

+++ tangos/testing/__init__.py
@@ -151,6 +151,8 @@ class SqlExecutionTracker(object):
         self._stacks.append("".join(traceback.format_list(traceback.extract_stack()[:-2])))
 
 def init_blank_db_for_testing(**init_kwargs):
+    DB_BACKEND = os.environ.get("DB_BACKEND", "sqlite")
+
     try:
         os.mkdir("test_dbs")
     except OSError:
@@ -160,12 +162,19 @@ def init_blank_db_for_testing(**init_kwargs):
 
     testing_db_name = init_kwargs.pop("testing_db_name", caller_fname)
 
-    db_name = "test_dbs/%s.db"%testing_db_name
-    try:
+    if DB_BACKEND == "sqlite":
+        db_name = f"test_dbs/%s.db"%testing_db_name
+        try:
+            os.remove(db_name)
+        except OSError:
+            pass
 
-        os.remove(db_name)
-    except OSError:
-        pass
+        core.init_db(f"sqlite:///{db_name}", **init_kwargs)
+    else:
+        user = os.environ.get("DB_USER")
+        password = os.environ.get("DB_PASSWORD")
+        port = os.environ.get("DB_PORT")
+        backend = DB_BACKEND.lower()
 
-    core.init_db("sqlite:///"+db_name,**init_kwargs)
-
+        url = f"{backend}://{user}:{password}@localhost:{port}/{testing_db_name}"
+        core.init_db(url, **init_kwargs)
